Return an error if Marketo reports an issue with bulk operations
Marketo may return an HTTP 200, along with an error in the
response. This commit rationalizes our error response handling for
bulk operations, turning any errors in the "success" response into a
Go error.

Additionally, we redefine our Error constants as go errors in
themselves, so we can use errors.Is to test the specific reasons
returned by the API.